backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
openai==1.27.0
pyautogen==0.2.27
python-dotenv==1.0.1
gunicorn==22.0.0

flake8==7.0.0
isort==5.13.2
black==24.4.2